Extract organization access CLI domain
